refactor: UserFeedback component
